* Fixed lots of warnings. * Fixed various build configuration issues. * Fixed some race conditions with dispoing GLControl objects on exit. There are still some rarer race conditions present for this issue that still need to be tracked down.

+++ Visualizer/PdbParser/PdbParser.cpp
@@ -27,7 +27,7 @@ void PdbParser::Open(String^ pdbPath)
 
 String^ PdbParser::GetFunctionName(UInt64 address)
 {
-	BSTR functionName = DumpSymbolWithRVA(pDiaSession, address, NULL);
+	BSTR functionName = DumpSymbolWithRVA(pDiaSession, (DWORD)address, NULL);
 	if (functionName != NULL)
 	{
 		String^ finalString = gcnew String((wchar_t*)functionName);
